Python 3 support
