+++ bindings/python/test.py
@@ -1,21 +1,23 @@
 #!/usr/bin/env python
-
+from __future__ import print_function
 # Capstone Python bindings, by Nguyen Anh Quynnh <aquynh@gmail.com>
 
 from capstone import *
 import binascii
+import sys
+_python3 = sys.version_info.major == 3
 
-X86_CODE16 = "\x8d\x4c\x32\x08\x01\xd8\x81\xc6\x34\x12\x00\x00"
-X86_CODE32 = "\x8d\x4c\x32\x08\x01\xd8\x81\xc6\x34\x12\x00\x00"
-X86_CODE64 = "\x55\x48\x8b\x05\xb8\x13\x00\x00"
-ARM_CODE = "\xED\xFF\xFF\xEB\x04\xe0\x2d\xe5\x00\x00\x00\x00\xe0\x83\x22\xe5\xf1\x02\x03\x0e\x00\x00\xa0\xe3\x02\x30\xc1\xe7\x00\x00\x53\xe3"
-ARM_CODE2 = "\x10\xf1\x10\xe7\x11\xf2\x31\xe7\xdc\xa1\x2e\xf3\xe8\x4e\x62\xf3"
-THUMB_CODE = "\x70\x47\xeb\x46\x83\xb0\xc9\x68"
-THUMB_CODE2 = "\x4f\xf0\x00\x01\xbd\xe8\x00\x88\xd1\xe8\x00\xf0"
-MIPS_CODE = "\x0C\x10\x00\x97\x00\x00\x00\x00\x24\x02\x00\x0c\x8f\xa2\x00\x00\x34\x21\x34\x56"
-MIPS_CODE2 = "\x56\x34\x21\x34\xc2\x17\x01\x00"
-ARM64_CODE = "\x21\x7c\x02\x9b\x21\x7c\x00\x53\x00\x40\x21\x4b\xe1\x0b\x40\xb9"
-PPC_CODE = "\x80\x20\x00\x00\x80\x3f\x00\x00\x10\x43\x23\x0e\xd0\x44\x00\x80\x4c\x43\x22\x02\x2d\x03\x00\x80\x7c\x43\x20\x14\x7c\x43\x20\x93\x4f\x20\x00\x21\x4c\xc8\x00\x21"
+X86_CODE16 = b"\x8d\x4c\x32\x08\x01\xd8\x81\xc6\x34\x12\x00\x00"
+X86_CODE32 = b"\x8d\x4c\x32\x08\x01\xd8\x81\xc6\x34\x12\x00\x00"
+X86_CODE64 = b"\x55\x48\x8b\x05\xb8\x13\x00\x00"
+ARM_CODE = b"\xED\xFF\xFF\xEB\x04\xe0\x2d\xe5\x00\x00\x00\x00\xe0\x83\x22\xe5\xf1\x02\x03\x0e\x00\x00\xa0\xe3\x02\x30\xc1\xe7\x00\x00\x53\xe3"
+ARM_CODE2 = b"\x10\xf1\x10\xe7\x11\xf2\x31\xe7\xdc\xa1\x2e\xf3\xe8\x4e\x62\xf3"
+THUMB_CODE = b"\x70\x47\xeb\x46\x83\xb0\xc9\x68"
+THUMB_CODE2 = b"\x4f\xf0\x00\x01\xbd\xe8\x00\x88\xd1\xe8\x00\xf0"
+MIPS_CODE = b"\x0C\x10\x00\x97\x00\x00\x00\x00\x24\x02\x00\x0c\x8f\xa2\x00\x00\x34\x21\x34\x56"
+MIPS_CODE2 = b"\x56\x34\x21\x34\xc2\x17\x01\x00"
+ARM64_CODE = b"\x21\x7c\x02\x9b\x21\x7c\x00\x53\x00\x40\x21\x4b\xe1\x0b\x40\xb9"
+PPC_CODE = b"\x80\x20\x00\x00\x80\x3f\x00\x00\x10\x43\x23\x0e\xd0\x44\x00\x80\x4c\x43\x22\x02\x2d\x03\x00\x80\x7c\x43\x20\x14\x7c\x43\x20\x93\x4f\x20\x00\x21\x4c\xc8\x00\x21"
 
 all_tests = (
         (CS_ARCH_X86, CS_MODE_16, X86_CODE16, "X86 16bit (Intel syntax)", 0),
@@ -27,34 +29,50 @@ all_tests = (
         (CS_ARCH_ARM, CS_MODE_ARM, ARM_CODE2, "ARM: Cortex-A15 + NEON", 0),
         (CS_ARCH_ARM, CS_MODE_THUMB, THUMB_CODE, "THUMB", 0),
         (CS_ARCH_MIPS, CS_MODE_32 + CS_MODE_BIG_ENDIAN, MIPS_CODE, "MIPS-32 (Big-endian)", 0),
-        (CS_ARCH_MIPS, CS_MODE_64+ CS_MODE_LITTLE_ENDIAN, MIPS_CODE2, "MIPS-64-EL (Little-endian)", 0),
+        (CS_ARCH_MIPS, CS_MODE_64 + CS_MODE_LITTLE_ENDIAN, MIPS_CODE2, "MIPS-64-EL (Little-endian)", 0),
         (CS_ARCH_ARM64, CS_MODE_ARM, ARM64_CODE, "ARM-64", 0),
         (CS_ARCH_PPC, CS_MODE_BIG_ENDIAN, PPC_CODE, "PPC-64", 0),
         (CS_ARCH_PPC, CS_MODE_BIG_ENDIAN, PPC_CODE, "PPC-64, print register with number only", CS_OPT_SYNTAX_NOREGNAME),
-        )
+)
 
 
 def to_hex(s):
-    return " ".join("0x" + "{0:x}".format(ord(c)).zfill(2) for c in s) # <-- Python 3 is OK
+    if _python3:
+        return " ".join("0x{0:02x}".format(c) for c in s)  # <-- Python 3 is OK
+    else:
+        return " ".join("0x{0:02x}".format(ord(c)) for c in s)
 
+def to_x(s):
+    from struct import pack
+    if not s: return '0'
+    x = pack(">q", s)
+    while x[0] in ('0', 0): x = x[1:]
+    return x
 
-### Test cs_disasm_quick()
+def to_x_32(s):
+    from struct import pack
+    if not s: return '0'
+    x = pack(">i", s)
+    while x[0] in ('0', 0): x = x[1:]
+    return x
+
+# ## Test cs_disasm_quick()
 def test_cs_disasm_quick():
-    for (arch, mode, code, comment, syntax) in all_tests:
+    for arch, mode, code, comment, syntax in all_tests:
         print('*' * 40)
-        print("Platform: %s" %comment)
+        print("Platform: %s" % comment)
         print("Disasm:"),
-        print to_hex(code)
+        print(to_hex(code))
         for insn in cs_disasm_quick(arch, mode, code, 0x1000):
-            print("0x%x:\t%s\t%s" %(insn.address, insn.mnemonic, insn.op_str))
-        print
+            print("0x%x:\t%s\t%s" % (insn.address, insn.mnemonic, insn.op_str))
+        print()
 
 
-### Test class Cs
+# ## Test class Cs
 def test_class():
-    for (arch, mode, code, comment, syntax) in all_tests:
+    for arch, mode, code, comment, syntax in all_tests:
         print('*' * 16)
-        print("Platform: %s" %comment)
+        print("Platform: %s" % comment)
         print("Code: %s" % to_hex(code))
         print("Disasm:")
 
@@ -65,16 +83,17 @@ def test_class():
                 md.syntax = syntax
 
             for insn in md.disasm(code, 0x1000):
-                #bytes = binascii.hexlify(insn.bytes)
-                #print("0x%x:\t%s\t%s\t// hex-code: %s" %(insn.address, insn.mnemonic, insn.op_str, bytes))
-                print("0x%x:\t%s\t%s" %(insn.address, insn.mnemonic, insn.op_str))
+                # bytes = binascii.hexlify(insn.bytes)
+                # print("0x%x:\t%s\t%s\t// hex-code: %s" %(insn.address, insn.mnemonic, insn.op_str, bytes))
+                print("0x%x:\t%s\t%s" % (insn.address, insn.mnemonic, insn.op_str))
 
             print("0x%x:" % (insn.address + insn.size))
-            print
+            print()
         except CsError as e:
-            print("ERROR: %s" %e)
+            print("ERROR: %s" % e)
 
 
-#test_cs_disasm_quick()
-#print "*" * 40
-test_class()
+# test_cs_disasm_quick()
+# print ("*" * 40)
+if __name__ == '__main__':
+    test_class()
